% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adm2.link.all}
\alias{adm2.link.all}
\title{WHO Administrative Level 2 Linkage}
\format{
A data frame with 1,095 rows and 8 columns:
\describe{
  \item{WHO_REGION}{WHO-designated region for the country.}
  \item{ISO.3.DIGIT.COUNTRY.CODE}{Three-letter ISO country code.}
  \item{ADM0_VIZ_NAME}{Administrative level 0 (country) name for visualization.}
  \item{ADM1_VIZ_NAME}{Administrative level 1 (first subnational division) name.}
  \item{ADM2_VIZ_NAME}{Administrative level 2 (second subnational division) name.}
  \item{GUID.LEVEL.0}{Global unique identifier for level 0 (country).}
  \item{GUID.LEVEL.1..}{Global unique identifier for level 1 (first subnational division).}
  \item{GLOBAL.UNIQUE.IDENTIFIER..}{Global unique identifier for level 2 (second subnational division).}
}
}
\usage{
adm2.link.all
}
\description{
A dataset linking administrative level 2 regions to their corresponding WHO region and country codes.
}
\keyword{datasets}
