% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHS.dataset.meta}
\alias{DHS.dataset.meta}
\title{DHS Dataset Metadata}
\format{
A data frame with 5,636 rows and 13 columns:
\describe{
  \item{FileFormat}{Format of the dataset file (e.g., Stata, SPSS, CSV).}
  \item{FileSize}{Size of the dataset file in megabytes.}
  \item{DatasetType}{Type of dataset/recode (e.g., Household, Individual, Children, Biomarker).}
  \item{SurveyNum}{Survey number assigned by DHS.}
  \item{SurveyId}{Unique ID of the survey associated with the dataset.}
  \item{FileType}{Type of file (e.g., Household Recode, Birth Recode, Men's Recode).}
  \item{FileDateLastModified}{Date when the dataset file was last modified.}
  \item{SurveyType}{Type of DHS survey (e.g., DHS, AIS, MIS).}
  \item{SurveyYearLabel}{Label describing the survey year.}
  \item{SurveyYear}{Year in which the DHS survey was conducted.}
  \item{DHS_CountryCode}{Unique country code assigned by DHS.}
  \item{FileName}{Name of the dataset file.}
  \item{CountryName}{Full name of the country associated with the dataset.}
}
}
\usage{
DHS.dataset.meta
}
\description{
A dataset containing metadata on recodes for DHS surveys supported by the app.
}
\keyword{datasets}
