% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHS_api_est}
\alias{DHS_api_est}
\title{DHS API Estimates}
\format{
A data frame with 28,110 rows and 7 columns:
\describe{
  \item{Country}{Name of the country where the survey was conducted.}
  \item{Country Code}{Three-letter ISO country code.}
  \item{Survey Year}{Year in which the DHS survey was conducted.}
  \item{DHS Standard ID}{Unique identifier for the DHS survey.}
  \item{Definition}{Definition of the health indicator being estimated.}
  \item{Estimate}{Estimated value of the indicator.}
  \item{By Variable Label}{Label describing any subgrouping (e.g., urban/rural, age group).}
}
}
\usage{
DHS_api_est
}
\description{
A dataset containing estimated health indicators from the DHS API, including country-level estimates and subgroup breakdowns.
}
\keyword{datasets}
