% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natl.WHO.shp}
\alias{natl.WHO.shp}
\title{WHO National-Level Shapefile}
\format{
A spatial data frame with 8 rows and 35 columns:
\describe{
  \item{WHO_REGION}{WHO-designated region for the country.}
  \item{ISO_3_CODE}{Three-letter ISO country code.}
  \item{ADM0_NAME}{Administrative level 0 (country) name.}
  \item{ADM0_CODE}{Administrative level 0 country code.}
  ...
}
}
\usage{
natl.WHO.shp
}
\description{
A spatial dataset containing national-level WHO boundaries for eight countries, with administrative codes and metadata.
}
\keyword{datasets}
