% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_tab_new}
\alias{ref_tab_new}
\title{Newly Added Indicators Reference Table}
\format{
A data frame with 134 rows and 13 columns:
\describe{
  \item{ID}{Unique identifier for the indicator.}
  \item{Description}{Short label for the indicator.}
  \item{Full_definition}{Detailed definition of the indicator.}
  \item{Topic}{General category or theme of the indicator.}
  \item{Chap_abbrev}{Abbreviation of the chapter where the indicator appears.}
  \item{IR}{Logical. Whether the indicator is available in the Individual Recode dataset.}
  \item{PR}{Logical. Whether the indicator is available in the Household Members dataset.}
  \item{KR}{Logical. Whether the indicator is available in the Children's Recode dataset.}
  \item{BR}{Logical. Whether the indicator is available in the Birth Recode dataset.}
  \item{HR}{Logical. Whether the indicator is available in the Household Recode dataset.}
  \item{MR}{Logical. Whether the indicator is available in the Men's Recode dataset.}
  \item{AR}{Logical. Whether the indicator is available in the AIDS Indicator Survey dataset.}
  \item{CR}{Logical. Whether the indicator is available in the Couple's Recode dataset.}
}
}
\usage{
ref_tab_new
}
\description{
A dataset containing newly added health and demographic indicators in the app after its initial release.
}
\keyword{datasets}
